Critical_Real/requirements.txt
streamlit>=1.10.0
pandas>=1.3.0
plotly>=5.3.0
python-dotenv>=0.19.0
requests>=2.26.0
folium>=0.12.0
geopandas>=0.10.0
shapely>=2.0.0
streamlit-folium>=0.12.0
